Issue #2326: Update blackList logic for pinnedTile
